refactor: create a component for pagination
